requirements.txt
# mechverse-bench requirements
# Python version: 3.11 (recommended)
openai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
Pillow>=10.0.0
